Move custom event dispatching from LayoutContainer, up to UI component.
* This enables the global pan/zoom gestures to continue to work even
  if the container has been moved outside of the viewport.
