+++ src/ui/RegisterComponentMixin.js
@@ -5,6 +5,13 @@
 
 export let RegisterComponentMixin = (superclass) => class extends superclass {
 
+  oninit(vnode) {
+    if(super.oninit) super.oninit(vnode);
+    if(vnode.attrs && vnode.attrs.registerComponentCallback) {
+      vnode.attrs.registerComponentCallback(this);
+    }
+  }
+
   oncreate(vnode) {
     if(super.oncreate) super.oncreate(vnode);
     vnode.dom.mithrilComponent = this;
