+++ src/ui/layout/CircosLayout.js
@@ -7,7 +7,7 @@ import {mix} from '../../../mixwith.js/src/mixwith';
 
 import {LayoutBase} from './LayoutBase';
 import {Bounds} from '../../model/Bounds';
-import {RegisterComponentMixin} from './RegisterComponentMixin';
+import {RegisterComponentMixin} from '../RegisterComponentMixin';
 
 const radians = degrees => degrees * Math.PI / 180;
 //const degrees = radians => radians * 180 / Math.PI; // TODO: remove if unused
