+++ src/ui/layout/LayoutContainer.js
@@ -5,16 +5,13 @@
   */
 import m from 'mithril';
 import PubSub from 'pubsub-js';
-import Hammer from 'hammerjs';
-import Hamster from 'hamsterjs';
-
 import {mix} from '../../../mixwith.js/src/mixwith';
 
 import {HorizontalLayout} from './HorizontalLayout';
 import {CircosLayout} from './CircosLayout';
 import {layout as layoutMsg, reset} from '../../topics';
 import {Bounds} from '../../model/Bounds';
-import {RegisterComponentMixin} from './RegisterComponentMixin';
+import {RegisterComponentMixin} from '../RegisterComponentMixin';
 
 
 export class LayoutContainer extends mix().with(RegisterComponentMixin) {
@@ -25,10 +22,19 @@ export class LayoutContainer extends mix().with(RegisterComponentMixin) {
    * mithril lifecycle method
    */
   oninit(vnode) {
+    super.oninit(vnode);
     this.appState = vnode.attrs.appState;
 
     PubSub.subscribe(layoutMsg, (msg, data) => this._onLayoutChange(msg, data));
     PubSub.subscribe(reset, (msg, data) => this._onReset(msg, data));
+
+    // create some regular expressions for faster dispatching of events
+    this._gestureRegex = {
+      pan:   new RegExp('^pan'),
+      pinch: new RegExp('^pinch'),
+      tap:   new RegExp('^tap'),
+      wheel: new RegExp('^wheel')
+    };
   }
 
   /**
@@ -37,7 +43,7 @@ export class LayoutContainer extends mix().with(RegisterComponentMixin) {
   oncreate(vnode) {
     super.oncreate(vnode);
     this.el = vnode.dom; // this is the outer m('div') from view()
-    this._setupEventHandlers(this.el);
+    //this._setupEventHandlers(this.el);
     this.bounds = new Bounds(this.el.getBoundingClientRect());
     this.contentBounds = new Bounds({
       left: 0,
@@ -53,14 +59,6 @@ export class LayoutContainer extends mix().with(RegisterComponentMixin) {
   onupdate(vnode) {
     console.assert(this.el === vnode.dom);
     this.bounds = new Bounds(vnode.dom.getBoundingClientRect());
-  }
-
-  /**
-   * mithril lifecycle method
-   */
-  onbeforeremove(vnode) {
-    super.onbeforeremove(vnode);
-    this._gestureEventsTeardown();
   }
 
   /**
@@ -81,63 +79,6 @@ export class LayoutContainer extends mix().with(RegisterComponentMixin) {
       :
       m(CircosLayout, {appState: this.appState, layoutBounds: this.bounds})
     ]);
-  }
-
-  /* private functions  */
-
-  _setupEventHandlers() {
-    //
-    // mouse wheel (hamsterjs)
-    //
-    let hamsterHandler = (evt) => {
-      // note: the hamster callback has additioanl parameters which are being
-      // ignored here (evt, delta, deltaX, deltaY)
-      // normalize the event so it is similar to the hammerjs gestures
-      evt.center = { x: evt.originalEvent.x, y: evt.originalEvent.y };
-      this._dispatchGestureEvt(evt);
-    };
-    let hamster = Hamster(this.el);
-    hamster.wheel(hamsterHandler);
-    //
-    // gestures (hammerjs)
-    //
-    let hammer = Hammer(this.el);
-    hammer.get('pan').set({ direction: Hammer.DIRECTION_ALL });
-    hammer.get('pinch').set({ enable: true });
-    let hammerHandler = (evt) => this._dispatchGestureEvt(evt);
-    let hammerEvents = 'panmove panend pinchmove pinchend tap';
-    hammer.on(hammerEvents, hammerHandler);
-    // create a callback for teardown of all these event handlers
-    this._gestureHandlerTeardown = () => {
-      hamster.unwheel(hamsterHandler);
-      hammer.off(hammerEvents, hammerHandler);
-    };
-    // create some regular expressions for faster dispatching of events
-    this._gestureRegex = {
-      pan:   new RegExp('^pan'),
-      pinch: new RegExp('^pinch'),
-      tap:   new RegExp('^tap'),
-      wheel: new RegExp('^wheel')
-    };
-  }
-
-  /**
-   * Custom dispatch of ui events. Layout elements like BioMap and
-   * CorrespondenceMap are visually overlapping, and so do not fit cleanly into
-   * the js event capture or bubbling phases. Query the dom at the events
-   * coordinates, and dispatch the event to child who
-   * a) intersects with this point
-   * b) wants to handle this event (it can decide whether to based on it's
-   * canvas scenegraph contents)
-   */
-  _dispatchGestureEvt(evt) {
-    let hitElements = document.elementsFromPoint(evt.center.x, evt.center.y);
-    let filtered = hitElements.filter( el => {
-      return el.mithrilComponent && el.mithrilComponent.handleGesture;
-    });
-    // dispatch event to all the mithril components, until one returns true;
-    // effectively the same as 'stopPropagation' on a normal event bubbling.
-    filtered.some( el => el.mithrilComponent.handleGesture(evt) );
   }
 
   /**
